refactor(node_api): extract test only used fns The api was only used in tests and can be moved to a separate struct.
