+++ sn_node/src/node/node_api/tests/mod.rs
@@ -8,15 +8,14 @@
 
 #![allow(dead_code, unused_imports)]
 
-use super::{Cmd, Comm, Dispatcher};
-
-use crate::comm::MsgEvent;
+use crate::comm::{Comm, MsgEvent};
 use crate::dbs::UsedSpace;
 use crate::node::{
-    cfg::create_test_max_capacity_and_root_storage, messages::WireMsgUtils,
-    node_api::event_channel, relocation_check, ChurnId, Error, Event, MembershipEvent, Node,
-    Proposal, RateLimits, Result as RoutingResult, RESOURCE_PROOF_DATA_SIZE,
-    RESOURCE_PROOF_DIFFICULTY,
+    cfg::create_test_max_capacity_and_root_storage,
+    messages::WireMsgUtils,
+    node_api::{dispatcher::Dispatcher, event_channel},
+    relocation_check, ChurnId, Cmd, Error, Event, MembershipEvent, Node, Proposal, RateLimits,
+    Result as RoutingResult, RESOURCE_PROOF_DATA_SIZE, RESOURCE_PROOF_DIFFICULTY,
 };
 
 use sn_interface::{
